Hide Auth controller from swagger documentation
